+++ verik-compiler/src/main/kotlin/io/verik/compiler/core/common/Core.kt
@@ -181,7 +181,7 @@ object Core {
         val F_time = CoreVkSystem.F_time
 
         val F_imported = CoreVkSpecial.F_imported
-        val F_sv_String = CoreVkSpecial.F_sv_String
+        val F_inject_String = CoreVkSpecial.F_inject_String
         val F_nc = CoreVkSpecial.F_nc
         val F_b = CoreVkSpecial.F_b
         val F_i = CoreVkSpecial.F_i